customize to our system

+++ kitti360scripts/viewer/kitti360Viewer3D.py
@@ -129,7 +129,7 @@ class Kitti360Viewer3D(object):
                 color[globalIds==uid] = self.getColor(instanceId)
             else:
                 color[globalIds==uid] = (96,96,96) # stuff objects in instance mode
-        color = color.astype(np.float)/255.0
+        color = color.astype(np.float32)/255.0
         return color
 
     def assignColorConfidence(self, confidence):
@@ -182,7 +182,7 @@ class Kitti360Viewer3D(object):
             color=np.vstack((data['red'], data['green'], data['blue'])).T
             pcd = open3d.geometry.PointCloud()
             pcd.points = open3d.utility.Vector3dVector(points)
-            pcd.colors = open3d.utility.Vector3dVector(color.astype(np.float)/255.)
+            pcd.colors = open3d.utility.Vector3dVector(color.astype(np.float32)/255.)
         
         # assign color
         if colorType=='semantic' or colorType=='instance':
